#2006: Organize and refine loader/document
* Segregate native parser code from SLIME parser code
* Document remaining usage of native parser
