+++ contributor/README.html
@@ -227,7 +227,7 @@ java.lang.IllegalArgumentException: Could not find option with name inspect.Path
 							<div>
 								One other potential goal, although jsapi is going away, would be to replace DOM in
 								<code>loader/browser/test/api.js</code> and
-								JaspiHtml in loader/api/old/jsh/html.js; this feeds
+								JsapiHtml in loader/api/old/jsh/html.js; this feeds
 								into loader/api/old/api.html.js as the <code>html</code> argument to ApiHtmlTests.
 							</div>
 							<div>
