+++ loader/document/module.fifty.ts
@@ -53,6 +53,15 @@ namespace slime.runtime.document {
 		xml?: boolean
 	}
 
+	(
+		function(
+			fifty: slime.fifty.test.Kit
+		) {
+			fifty.tests.exports = fifty.test.Parent();
+		}
+	//@ts-ignore
+	)(fifty);
+
 	export namespace document {
 		export interface Exports {
 			codec: {
@@ -62,8 +71,10 @@ namespace slime.runtime.document {
 				string: (settings: Settings) => (string: string) => slime.runtime.document.Document
 			}
 			removeWhitespaceTextNodes: slime.$api.fp.Transform<slime.runtime.document.Document>
+		}
+
+		export interface Exports {
 			prettify: (p: { indent: string }) => slime.$api.fp.Transform<slime.runtime.document.Document>
-			element: (p: slime.runtime.document.Document) => slime.runtime.document.Element
 		}
 
 		(
@@ -72,7 +83,7 @@ namespace slime.runtime.document {
 			) {
 				var subject = test.subject;
 
-				fifty.tests.Document = function() {
+				fifty.tests.exports.Document = function() {
 					var before = "<root><child/><child/><child/></root>";
 					var document = subject.Document.codec.string.decode(before);
 					var pretty = subject.Document.prettify({
@@ -94,6 +105,10 @@ namespace slime.runtime.document {
 			}
 		//@ts-ignore
 		)(fifty);
+
+		export interface Exports {
+			element: (p: slime.runtime.document.Document) => slime.runtime.document.Element
+		}
 	}
 
 	export interface Exports {
@@ -243,108 +258,19 @@ namespace slime.runtime.document {
 		Element: element.Exports
 	}
 
-	export type Script = slime.loader.Script<Context | void,Exports>
-}
-
-/**
- * The SLIME document parser is a parser and serializer that handles HTML (and XHTML/XML) documents, converting them to a
- * tree structure (which is somewhat analogous to a DOM tree).
- *
- * The produced tree conforms to the HTML DOM specification, so (for example) whitespace may be altered when parsing.
- */
-namespace slime.runtime.document.old {
-	export interface Context {
-		/**
-		 * Only used when using the JSoup implementation.
-		 */
-		$slime?: slime.jsh.plugin.$slime
-	}
-
-	export interface Exports {
-		load: any
-	}
-
-	export type Script = slime.loader.Script<Context,Exports>;
-
-	export namespace test {
-		export const subject = (function(fifty: slime.fifty.test.Kit) {
-			var load: Script = fifty.$loader.script("module.js");
-
-			var isBrowser = Boolean(fifty.global.window);
-			var isJsh = Boolean(fifty.global.jsh);
-			var api: Exports = load({
-				$slime: (isJsh) ? fifty.global.jsh.unit.$slime : void(0)
-			});
-			return api;
-		//@ts-ignore
-		})(fifty);
-	}
-
 	(
 		function(
 			fifty: slime.fifty.test.Kit
 		) {
-			var api = test.subject;
-
-			function testsFor1(page,verify,withComments) {
-				var offset = (withComments) ? 1 : 0
-				verify(page).children.length.is(2 + offset);
-				verify(page).children.get(0 + offset).doctype.is.type("object");
-				verify(page).children.get(0 + offset).doctype.name.is.type("string");
-				verify(page).children.get(0 + offset).doctype.name.is("html");
-
-				verify(page).children.get(1 + offset).element.is.type("object");
-				verify(page).document.element.is.type("object");
-				verify(page).document.element.element.name.is("html");
-
-				verify(page).document.element.element.attributes.get("class").is("class1");
-				verify(page).document.element.element.attributes.get("foo").is(null);
-			}
-
-			fifty.tests.old = function() {
-				fifty.run(function files() {
-					//	This is not really a test of this module
-					var resource = fifty.$loader.get("test/data/1.html");
-					fifty.verify(resource).is.type("object");
-					var missing = fifty.$loader.get("test/data/foo.html");
-					fifty.verify(missing).is.type("null");
-				});
-
-				fifty.run(function parse() {
-					var page = api.load({
-						loader: fifty.$loader,
-						path: "test/data/1.html"
-					});
-					testsFor1(page,fifty.verify,true);
-				});
-
-				fifty.run(function codec() {
-					var page = api.load({
-						loader: fifty.$loader,
-						path: "test/data/1.html"
-					});
-					var string: string = page.serialize();
-					var reparsed = api.load({
-						string: string
-					});
-					testsFor1(reparsed,fifty.verify,Boolean(fifty.global.jsh));
-					//	TODO	below does not work in JSoup 1.12.1, which messes up some pieces of whitespace
-					//	TODO	below does not work in browser, either
-					//	See https://stackoverflow.com/questions/57394776/why-does-domparser-alter-whitespace
-					if (false) {
-						fifty.verify(string).is(fifty.$loader.get("test/data/1.html").read(String));
-					}
-				});
-			}
-
 			fifty.tests.suite = function() {
-				fifty.run(fifty.tests.Document);
-
 				fifty.load("source.fifty.ts");
+				fifty.run(fifty.tests.exports);
 
-				fifty.run(fifty.tests.old);
+				fifty.load("old.fifty.ts");
 			}
 		}
 	//@ts-ignore
 	)(fifty);
+
+	export type Script = slime.loader.Script<Context | void,Exports>
 }
